Abstracting and moving things to trait. Skip over non-local images. Split the filepath and filename.
